fix: Tira a responsabilidade do model missões e passa para o DTO nas requisições

+++ src/main/java/com/kava/s/CadastroNinja/missao/service/MissaoService.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class MissaoService {
@@ -28,23 +29,30 @@ public class MissaoService {
     }
 
     //GET -- Manda uma requisição para listar todos as missões
-    public List<MissaoModel> listarMissoes(){
-        return missaoRepository.findAll();
+    public List<MissaoDTO> listarMissoes(){
+        List<MissaoModel> missaoModel = missaoRepository.findAll();
+
+        return missaoModel.stream()
+                .map(missaoMapper::map)
+                .collect(Collectors.toList());
     }
 
     //GET -- Manda uma requisição para listar missao por ID
-    public MissaoModel listarMissaoId(Long id){
+    public MissaoDTO listarMissaoId(Long id){
         Optional<MissaoModel> caixaMissao = missaoRepository.findById(id);
 
-        return caixaMissao.orElse(null);
+        return caixaMissao.map(missaoMapper::map).orElse(null);
     }
 
     //PUT -- Manda uma requisição para Alterar uma Missao
-    public MissaoModel alterarMissao(Long id, MissaoModel missaoAtualizada){
-        if(missaoRepository.existsById(id)){
-            missaoAtualizada.setId(id);
+    public MissaoDTO alterarMissao(Long id, MissaoDTO missaoDTO){
+        Optional<MissaoModel> caixaMissao = missaoRepository.findById(id);
 
-            return missaoRepository.save(missaoAtualizada);
+        if(caixaMissao.isPresent()){
+            MissaoModel missao = missaoMapper.map(missaoDTO);
+            missao.setId(id);
+
+            return missaoMapper.map(missao);
         }
 
         return null;
@@ -52,6 +60,10 @@ public class MissaoService {
 
     //DELETE -- Manda uma requisição para deletar uma Missão
     public void deletarMissao(Long id){
-        missaoRepository.deleteById(id);
+        Optional<MissaoModel> caixaMissao = missaoRepository.findById(id);
+
+        if(caixaMissao.isPresent()){
+            missaoRepository.deleteById(id);
+        }
     }
 }
